# [db] Reset only if Snapshots exceed size of 10